fix something in revenue admin
