+++ security/src/main/java/vn/iotstar/security/service/ProductService.java
@@ -8,6 +8,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import vn.iotstar.security.model.Category;
 import vn.iotstar.security.model.Product;
+import vn.iotstar.security.model.ProductOrder;
 import vn.iotstar.security.model.Shop;
 
 
@@ -45,4 +46,6 @@ public interface ProductService {
 	List<Product> getProductsSoldMoreThan10();
 	
 	public Boolean deleteProductByCategory(Category category);
+	
+	String getTotalSoldProduct(List<ProductOrder> allDeliveredOrders);
 }
