+++ security/src/main/java/vn/iotstar/security/service/impl/OrderServiceImpl.java
@@ -1,7 +1,12 @@
 package vn.iotstar.security.service.impl;
 
+import java.text.DecimalFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.time.Month;
+import java.util.Comparator;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -46,6 +51,8 @@ public class OrderServiceImpl implements OrderService {
     private FileStorageService fileStorageService;
     @Autowired
     private ReviewRepository reviewRepository;
+    @Autowired
+    private ShopServiceImpl shopService;
     @Override
     public void saveOrder(Integer userId, OrderRequest orderRequest) throws Exception {
         List<Cart> carts = cartRepository.findByUserId(userId);
@@ -242,5 +249,56 @@ public class OrderServiceImpl implements OrderService {
 
         return monthlyRevenueMap;
     }
+
+	@Override
+	public String getTotalRevenue() {
+		double totalRevenue = 0;
+		List<Shop> listShop = shopService.getAll();
+	    if (!listShop.isEmpty()) {
+	        for (Shop shop : listShop) {
+	            totalRevenue += shop.getRevenue();
+	        }
+	    }
+	    
+	    DecimalFormat df = new DecimalFormat("#,##0");
+	    String formattedTotalRevenue = df.format(totalRevenue);
+	    
+	    return formattedTotalRevenue;
+	}
+
+	@Override
+	public Map<String, Double> getMonthlyRevenue(List<ProductOrder> listDeliveredProductOrder, int year) {
+		// Use TreeMap with custom Comparator (corrected)
+        Map<String, Double> monthlyRevenueMap = new TreeMap<>(new Comparator<String>() {
+            @Override
+            public int compare(String monthYear1, String monthYear2) {
+                try {
+                    SimpleDateFormat format = new SimpleDateFormat("MMMM yyyy");
+                    Date date1 = format.parse(monthYear1);
+                    Date date2 = format.parse(monthYear2); 
+                    return date1.compareTo(date2);
+                } catch (ParseException e) {
+                    e.printStackTrace();
+                    return 0;
+                }
+            }
+        });
+
+        for (ProductOrder productOrder : listDeliveredProductOrder) {
+            LocalDate orderDate = productOrder.getOrderDate();
+            if (orderDate.getYear() == year) {
+                String monthYear = orderDate.getMonth().toString() + " " + orderDate.getYear();
+                monthlyRevenueMap.put(monthYear, monthlyRevenueMap.getOrDefault(monthYear, 0.0) + productOrder.getPrice());
+            }
+        }
+
+        for (Month month : Month.values()) {
+            String monthYear = month.toString() + " " + String.valueOf(year);
+            monthlyRevenueMap.putIfAbsent(monthYear, 0.0);
+        }
+        
+        return monthlyRevenueMap;
+	}
+	
 
 }
