+++ security/src/main/java/vn/iotstar/security/service/impl/ProductServiceImpl.java
@@ -19,6 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 import jakarta.transaction.Transactional;
 import vn.iotstar.security.model.Category;
 import vn.iotstar.security.model.Product;
+import vn.iotstar.security.model.ProductOrder;
 import vn.iotstar.security.model.Shop;
 import vn.iotstar.security.repository.CartRepository;
 import vn.iotstar.security.repository.CategoryRepository;
@@ -229,6 +230,15 @@ public class ProductServiceImpl implements ProductService {
 			return true;
 		}
 		return false;
+	}
+
+	@Override
+	public String getTotalSoldProduct(List<ProductOrder> allDeliveredOrders) {
+		int totalProduct = 0;
+		for (ProductOrder productOrder : allDeliveredOrders) {
+            totalProduct += productOrder.getQuantity();
+        }
+		return String.valueOf(totalProduct);
 	}
 
 }
