+++ security/src/main/resources/templates/admin/index.html
@@ -80,7 +80,7 @@
 						<div class="card card-sh">
 							<div class="card-body text-center text-primary">
 								<i class="fa-solid fa-square-plus fa-3x"></i>
-								<h4>Revuenue</h4>
+								<h4>Revenue</h4>
 							</div>
 						</div>
 					</a>
